requirements.txt
streamlit>=1.0.0
rembg
torch
torchvision
Pillow>=8.0.0
opencv-python>=4.5.0
numpy>=1.19.0
python-dotenv
openai
requests
onnxruntime
tensorflow>=2.4.0
matplotlib>=3.3.0
panda
